refactor: move to Infrastructure the IO classes from Command module
