+++ src/php/Command/Application/CommandExceptionWriter.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Phel\Command\Application;
 
-use Phel\Command\Domain\Shared\CommandExceptionWriterInterface;
-use Phel\Command\Domain\Shared\ErrorLog\ErrorLogInterface;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
+use Phel\Command\Domain\CommandExceptionWriterInterface;
+use Phel\Command\Domain\ErrorLogInterface;
+use Phel\Command\Domain\Exceptions\ExceptionPrinterInterface;
 use Phel\Compiler\Domain\Exceptions\AbstractLocatedException;
 use Phel\Compiler\Domain\Parser\ReadModel\CodeSnippet;
 use Symfony\Component\Console\Output\OutputInterface;
