+++ src/php/Command/Application/TextExceptionPrinter.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Phel\Command\Application;
 
-use Phel\Command\Domain\Shared\ErrorLog\ErrorLogInterface;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionArgsPrinterInterface;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
-use Phel\Command\Domain\Shared\Exceptions\Extractor\FilePositionExtractorInterface;
+use Phel\Command\Domain\ErrorLogInterface;
+use Phel\Command\Domain\Exceptions\ExceptionArgsPrinterInterface;
+use Phel\Command\Domain\Exceptions\ExceptionPrinterInterface;
+use Phel\Command\Domain\Exceptions\Extractor\FilePositionExtractorInterface;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\MungeInterface;
 use Phel\Compiler\Domain\Exceptions\AbstractLocatedException;
 use Phel\Compiler\Domain\Parser\ReadModel\CodeSnippet;
