+++ src/php/Command/CommandFactory.php
@@ -9,16 +9,16 @@ use Gacela\Framework\Config\ConfigReader\PhpConfigReader;
 use Phel\Command\Application\CommandExceptionWriter;
 use Phel\Command\Application\DirectoryFinder;
 use Phel\Command\Application\TextExceptionPrinter;
-use Phel\Command\Domain\Finder\ComposerVendorDirectoriesFinder;
+use Phel\Command\Domain\CommandExceptionWriterInterface;
+use Phel\Command\Domain\Exceptions\ExceptionArgsPrinter;
+use Phel\Command\Domain\Exceptions\ExceptionPrinterInterface;
+use Phel\Command\Domain\Exceptions\Extractor\FilePositionExtractor;
 use Phel\Command\Domain\Finder\DirectoryFinderInterface;
 use Phel\Command\Domain\Finder\VendorDirectoriesFinderInterface;
-use Phel\Command\Domain\Shared\CommandExceptionWriterInterface;
-use Phel\Command\Domain\Shared\ErrorLog\ErrorLog;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionArgsPrinter;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
-use Phel\Command\Domain\Shared\Exceptions\Extractor\FilePositionExtractor;
-use Phel\Command\Domain\Shared\Exceptions\Extractor\SourceMapExtractor;
+use Phel\Command\Infrastructure\ComposerVendorDirectoriesFinder;
+use Phel\Command\Infrastructure\ErrorLog;
 use Phel\Command\Infrastructure\ExceptionHandler;
+use Phel\Command\Infrastructure\SourceMapExtractor;
 use Phel\Compiler\Application\Munge;
 use Phel\Printer\Printer;
 use Phel\Run\Domain\Repl\ColorStyle;
