+++ src/php/Command/Infrastructure/ComposerVendorDirectoriesFinder.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Phel\Command\Domain\Finder;
+namespace Phel\Command\Infrastructure;
 
 use Gacela\Framework\DocBlockResolverAwareTrait;
 use Phel\Command\CommandFacade;
+use Phel\Command\Domain\Finder\VendorDirectoriesFinderInterface;
 use Phel\Config\PhelConfig;
 use Phel\Phel;
 use RuntimeException;
