+++ tests/php/Integration/Run/Command/Repl/ReplCommandTest.php
@@ -8,11 +8,11 @@ use Gacela\Framework\Bootstrap\GacelaConfig;
 use Gacela\Framework\ClassResolver\GlobalInstance\AnonymousGlobal;
 use Gacela\Framework\Gacela;
 use Generator;
-use Phel\Command\Domain\Shared\ErrorLog\ErrorLogInterface;
-use Phel\Command\Domain\Shared\Exceptions\ExceptionArgsPrinter;
-use Phel\Command\Domain\Shared\Exceptions\Extractor\FilePositionExtractor;
-use Phel\Command\Domain\Shared\Exceptions\Extractor\SourceMapExtractor;
-use Phel\Command\Domain\Shared\Exceptions\TextExceptionPrinter;
+use Phel\Command\Application\TextExceptionPrinter;
+use Phel\Command\Domain\ErrorLogInterface;
+use Phel\Command\Domain\Exceptions\ExceptionArgsPrinter;
+use Phel\Command\Domain\Exceptions\Extractor\FilePositionExtractor;
+use Phel\Command\Infrastructure\SourceMapExtractor;
 use Phel\Compiler\Application\Munge;
 use Phel\Printer\Printer;
 use Phel\Printer\PrinterInterface;
