+++ tests/php/Integration/Run/Command/Repl/ReplTestIo.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhelTest\Integration\Run\Command\Repl;
 
-use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
+use Phel\Command\Domain\Exceptions\ExceptionPrinterInterface;
 use Phel\Compiler\Domain\Exceptions\AbstractLocatedException;
 use Phel\Compiler\Domain\Parser\ReadModel\CodeSnippet;
 use Phel\Run\Domain\Repl\ReplCommandIoInterface;
